split radarr into more files
