+++ radarr/command.go
@@ -1,0 +1,37 @@
+package radarr
+
+import (
+	"encoding/json"
+	"fmt"
+)
+
+// GetCommands returns all available Radarr commands.
+func (r *Radarr) GetCommands() ([]*CommandResponse, error) {
+	var output []*CommandResponse
+
+	if err := r.GetInto("v3/command", nil, &output); err != nil {
+		return nil, fmt.Errorf("api.Get(command): %w", err)
+	}
+
+	return output, nil
+}
+
+// SendCommand sends a command to Radarr.
+func (r *Radarr) SendCommand(cmd *CommandRequest) (*CommandResponse, error) {
+	var output CommandResponse
+
+	if cmd == nil || cmd.Name == "" {
+		return &output, nil
+	}
+
+	body, err := json.Marshal(cmd)
+	if err != nil {
+		return nil, fmt.Errorf("json.Marshal(cmd): %w", err)
+	}
+
+	if err := r.PostInto("v3/command", nil, body, &output); err != nil {
+		return nil, fmt.Errorf("api.Post(command): %w", err)
+	}
+
+	return &output, nil
+}
