+++ radarr/system.go
@@ -1,0 +1,31 @@
+package radarr
+
+import (
+	"fmt"
+
+	"golift.io/starr"
+)
+
+// GetSystemStatus returns system status.
+func (r *Radarr) GetSystemStatus() (*SystemStatus, error) {
+	var status SystemStatus
+
+	err := r.GetInto("v3/system/status", nil, &status)
+	if err != nil {
+		return nil, fmt.Errorf("api.Get(system/status): %w", err)
+	}
+
+	return &status, nil
+}
+
+// GetBackupFiles returns all available Radarr backup files.
+// Use GetBody to download a file using BackupFile.Path.
+func (r *Radarr) GetBackupFiles() ([]*starr.BackupFile, error) {
+	var output []*starr.BackupFile
+
+	if err := r.GetInto("v3/system/backup", nil, &output); err != nil {
+		return nil, fmt.Errorf("api.Get(system/backup): %w", err)
+	}
+
+	return output, nil
+}
